+++ .gitignore
@@ -14,6 +14,7 @@
 
 # Project-local glide cache, RE: https://github.com/Masterminds/glide/issues/736
 .glide/
+.idea/
 ydict
 .DS_Store
 .env
